+++ internal/platform/handler.go
@@ -20,14 +20,6 @@ func NewPlatformHandler(db DBTX) Handler {
 	}
 }
 
-type Handler interface {
-	CreatePlatform(w http.ResponseWriter, r *http.Request)
-	GetPlatforms(w http.ResponseWriter, r *http.Request)
-	GetPlatform(w http.ResponseWriter, r *http.Request)
-	UpdatePlatform(w http.ResponseWriter, r *http.Request)
-	DeletePlatform(w http.ResponseWriter, r *http.Request)
-}
-
 type platformHandler struct {
 	queries Querier
 }
